added a timer example
In this example we show a Feedback that ignores the state changes and simply publishes something to the system based on a side affect that is not created from an event the user initiates. In this case a timer that is used to update the UI on a regular basis. This same pattern could be used for changes coming from another system, a persistance layer change, network reachability and so on.